+++ planning/autoware_diffusion_planner/src/preprocessing/lane_segments.cpp
@@ -115,8 +115,8 @@ std::pair<std::vector<float>, std::vector<float>> LaneSegmentContext::get_route_
 
 std::pair<std::vector<float>, std::vector<float>> LaneSegmentContext::get_lane_segments(
   const Eigen::Matrix4d & transform_matrix,
-  const std::map<lanelet::Id, TrafficSignalStamped> & traffic_light_id_map, const float center_x,
-  const float center_y, const int64_t m) const
+  const std::map<lanelet::Id, TrafficSignalStamped> & traffic_light_id_map, const double center_x,
+  const double center_y, const int64_t m) const
 {
   if (map_lane_segments_matrix_.rows() != SEGMENT_POINT_DIM || m <= 0) {
     throw std::invalid_argument(
